backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Proxmox API
proxmoxer==2.0.1
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Validation
marshmallow==3.20.1
jsonschema==4.20.0

# Subprocess management
python-terraform==0.10.1

# Git operations
GitPython==3.1.40

# Utilities
PyYAML==6.0.1
python-dateutil==2.8.2

# Security
werkzeug==3.0.1

# Logging
colorlog==6.8.0
